+++ src/cqterrain/floor/__init__.py
@@ -2,3 +2,4 @@ from .BaseFloor import BaseFloor
 from .RoundBrickFloor import RoundBrickFloor
 from .WoodFloor import WoodFloor
 from .TileGenerator import TileGenerator
+from .ModPattern import ModPattern