+++ src/cqterrain/minibase/__init__.py
@@ -14,7 +14,9 @@ from .slot_wood import slot_wood
 from .hexagon import hexagon
 from .hexagon_uneven import hexagon_uneven
 from .hexagon_wood import hexagon_wood
-from.RuinStoneBase import RuinStoneBase
+from .RuinStoneBase import RuinStoneBase
+from .PointGridBase import PointGridBase
+from .BasaltBase import BasaltBase
 
 from .base_irregular import base_irregular 
 from .circle_irregular import circle_irregular